requirements.txt
solana
python-telegram-bot==20.3
asyncio